Add files via upload
The jupyter notebook in conda environment showcases the various data visualisations based on the findings. I have used seaborn and matplotlib libraries to create these visualizations. I have also used a wordcloud to visualize the wide range of genres that is curated by Netflix for its wide customer base.

My main focus in this project was to gain insights on the genre, actors, directors. In future, I also plan to explore the ratings of the movies and tv shows and what type of content majority of subscribers tune in to.